--- app/views/wallet/ContactsBookView/ContactBox.d.ts
@@ -1,6 +1,0 @@
-export interface ContactBoxProps {
-  assignedAddress: string;
-  abbreviation: string;
-  alias: string;
-  onEdit: () => unknown;
-}
